Fix Escape and some control character processing.
The Escape key was being translated by the system
into a cancel command that was not being handled,
causing certain terminal applications to not
process it correctly.

Fixed the “remap backquote to Escape” preference
as well; if this is set, the backquote key sends
the same control character as Escape.

Also, the code for handling raw control characters
is now more flexible, supporting a broader range
of visible character translations (e.g. allowing
upper-case characters to be combined with control
instead of requiring them to be lowercase).