Simplify line buffers; begin port of “cell” API.
The TerminalLine_Object API has been simplified to
reduce dependencies on its internal buffers.  It
is preferable to access lines as Core Foundation
strings or by using methods that make it clearer
what the caller actually requires.

Some of the direct dependencies on UniChar buffers
that were in the Terminal module have been updated
to use CFStringRef and CFRange only.  Namely, the
terminal line range APIs (used for text selection)
no longer depend on internal buffers directly.

Of course, an index, range or composed character
sequence is not sufficient for terminals, which
rely on columns (or “cells”).  The columns that a
string occupies can be non-trivial to compute and
this requires API support.

Therefore, separately, substantial work has been
done to add the ability to compute these columns
(or “cells”).  This commit begins to integrate a
portion of those improvements, starting with the
new StringUtilities_Cell type (wrapping any value
that is meant to refer to a terminal column to
prevent accidental translation).  There are also
APIs meant to establish intent when translating
from columns to/from ranges in a string.

This is not being integrated all at once to avoid
requiring too many simultaneous changes.  (Even
new APIs may have temporary placeholders to ease
porting.)