Upgrade insert/delete code; fix corner cases.
The code to insert blanks or delete text ranges
is now implemented in terms of cells, which will
make it more straightforward to swap to a new
internal buffer type (needed to handle complex
Unicode correctly).  This is now also located in
the Terminal Line module, and the Terminal module
has been updated to use it.

In the process of testing this, fixed some issues
with text if the affected lines were double-size
or relying on 80/132-column mode switches.