Migrate Server Browser to SwiftUI.
This simplifies the interface considerably
(several hundred lines of code saved), and
updates the look and feel somewhat.  Since
table views are not well supported yet in
SwiftUI, an alternate list appearance was
used.

Field validation is fundamentally similar
to before but this is easier to implement
and alerts no longer appear.  Instead, any
error messages simply remain displayed in
the panel until they are resolved.

Layout constraints were adopted to make it
easier to properly adjust the window when
the list of services appears.

This also fixes an issue with completing
the background task to resolve host names.