Fix graphics-mode issues; expand VT52 support.
The per-character-translation function has been
updated to support Unicode and assume substrings
instead of single characters.  It is now also
aware of cell boundaries (as opposed to making
wrong assumptions about character indices).  The
cell-aware option is disabled for now, like all
other similar methods with a cell-aware option,
since every piece of the text system has to be
upgraded at the same time for consistency.

Several graphics characters that were not being
handled correctly with the new renderer are now
fixed.  In addition, graphics mode is now entered
and exited correctly in “VT52 mode”, and all the
graphics characters that are different in VT52
are now correctly remapped and displayed.  This
means that MacTerm finally supports VT52 graphics
correctly!