Fix auto-Copy to work for extended selections.
If “Automatically Copy selected text” is enabled
for a terminal, this once again works for text
that is selected with the mouse or extended by
the keyboard.  (This was broken when moving from
Carbon to Cocoa.  The new terminal did respect
auto-Copy for selections that were made with
double-clicks or triple-clicks but not other
mouse or keyboard actions.)