Expand DEC private modes (mostly placeholders).
Implemented XTerm modes that rely on saving
and restoring the cursor, such as mode 1048.

Added dozens of placeholders for all known DEC
private modes that XTerm supports, where the
default behavior is to log any set or reset of
those modes.  This will make it easier to add
implementations of those in future commits,
and easier to notice when a program is trying
to use them.