Implement “bracketed Paste” mode.
DEC private mode 2004 is now supported, also known
as “bracketed Paste”.  This causes MacTerm to emit
special terminal sequences just before and after
text has been sent to the session using Paste.  It
is also used for an equivalent user action, like
drag-and-drop of a chunk of text.

Newer versions of some shells and editors tend to
support this mode; e.g. a shell may allow the text
to be inserted and wait for confirmation instead
of trying to execute every line.

Since the Multi-Line Paste warning was primarily
designed to prevent disasters from sending text to
programs in unintended ways, it is now implicitly
deactivated while in “bracketed Paste” mode.  The
assumption is that the program currently running
in the terminal will handle everything reasonably.

The mode can be turned on/off at different times,
usually coinciding with starting a new program or
seeing a new shell prompt.  To make it easier to
see what may happen when Paste is used, the main
menu and contextual menu now indicate the state of
this protected mode (below the Paste command).

A low-level setting has been added to disable this
new “bracketed Paste” support entirely, although
it is now respected by default.

The release notes have been updated to describe
all these features in some detail.

Also, updated the release notes to mention other
recent Python-related changes.