Improve focus-follows-mouse behavior.
This now checks for certain parent-window
scenarios in focus-follows-mouse mode, to
prevent mouse movements over a main window
from causing child windows to disappear
(e.g. Find panel or remote server panel).
It is still possible to switch focus to an
entirely different window while a child is
open on the previous window.