Update startup OS version logic.
This fixes the warning that would appear in the
log on Monterey, and updates the logic to have
the correct behavior for more recent OSes.

Note that in many scenarios the Finder will
simply fail to launch on older systems so some
of this logic may only apply to command-line
launches now.