Suppress warnings about set-but-unused variables.
Added compiler flag "-Wno-unused-but-set-variable"
to prevent new default behavior that warns when a
variable’s value is not used before it changes.
This is annoying because all variables are given
initial values (to make sure they are never left
uninitialized/random even after code changes),
and it is worse to have to make them uninitialized
just to satisfy a warning.