Fix focus-follows-mouse when selecting menu items.
Focus-follows-mouse no longer takes effect
immediately after selecting a menu item
manually with the mouse.  This ensures that
the menu command applies to the intended
target; otherwise it was possible for the
focus to switch to whatever window happened
to be under the selected menu item (as soon
as the menu disappeared, the focus could
switch, then the command would take effect).