Fix live resize for non-terminal windows.
The floating window with terminal screen dimensions
or font size now only appears if the terminal view
is in an actual terminal window.  In particular, if
the terminal view is being used for something else
(like a preview in a preferences pane), resizing
its window no longer shows the extra info window.

Added "isUsingInfoBubble" method so that terminal
views can easily tell when to show/hide or update
the information window during resizes.  Also fixed
the relevant overridden methods to call "super".