Revert Big Sur style alerts; more tweaks.
Alerts have returned to a wider, left-oriented
layout, ending the long experiment to try to
resemble the system’s more recent narrow style.

This was given a fair trial over a very long
period of time but frankly it is quite clear
that the traditional arrangement is superior in
every possible way.

Experiments with button colors have also been
tweaked, improving legibility over the default
reds that the system would otherwise choose for
“destructive” actions.