Fix “Make Text as Big as Possible”.
This feature was previously broken when
migrating from Carbon to Cocoa but has
been reimplemented to work with the new
terminal implementation.  The terminal
screen’s column and row counts remain
unchanged but the font is changed to be
as big as it can (where the window will
still fit within usable screen space).