Fix “Find Cursor”; improve animation.
Reimplemented the “Find Cursor” command
that worked in the old Carbon version of
terminal windows but had not yet been
fixed for the new Cocoa terminal windows.

Simplified the animation routines to rely
directly on window animation capabilities,
which in recent versions of the OS are now
quite smooth.  (Removed some flexibility
in the internal animation API to make this
easier to port over.)